Hook up unsubscribe feature to back-end
